+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Upcoming
 
+# 0.3.0
+
+This release corresponds with Convex npm release 0.13.0.
+
+- `mutation()`, `action()`, and `query()` now take two arguments: the function
+  name and an (optional) arguments dictionary. See
+  https://blog.convex.dev/announcing-convex-0-13-0/ for more about this change.
+
+  If you need to communicate with a Convex backend with functions with
+  positional arguments instead of a single arguments object, _stay on version
+  0.2.0_.
+
 # 0.2.0
 
 - Both python `int` and `float` will coerce to a `number` type in Convex's JS